Added parse time string function in timer
