+++ display.py
@@ -1,41 +1,12 @@
-import curses
-import curses.panel
 import sys
 import os
 import csv
+from user_interface import UserInterface
 from select import select
-from collections import OrderedDict
 
 import time
 
 import timer
-
-
-class UserInterface:
-    def __init__(self):
-        self.stdscr = curses.initscr()
-        curses.noecho()
-        curses.cbreak()
-        curses.curs_set(0)
-        self.stdscr.keypad(1)
-
-        self.win = curses.newwin(20, 70, 0, 0)
-        self.win.border(0)
-
-        self.win.addstr(1, 30, "Tracker")
-        self.win.addstr(2, 30, "=======")
-        self.win.addstr(4, 1, "Press 'p' to pause/unpause or 'q' to quit.")
-
-    def refresh(self):
-        curses.panel.update_panels()
-        self.win.refresh()
-
-    def quit_ui(self):
-        curses.nocbreak()
-        self.stdscr.keypad(0)
-        curses.curs_set(1)
-        curses.echo()
-        curses.endwin()
 
 
 class Feeder:
@@ -85,14 +56,16 @@ class Feeder:
                     elif line.strip() == "s":
                         start = time.ctime(self.first_start)
                         finish = time.ctime()
-                        logged = timer.get_time_string(self.start, self.recorded_time)
+                        if self.paused:
+                            logged = self.recorded_time
+                        else:
+                            logged = timer.get_time_string(self.start, self.recorded_time)
                         current_row = {"Start": start, "Finish": finish, "Logged": logged}
                         lines = []
                         try:
                             with open(f"./{self.task}.csv", "r") as csvfile:
                                 taskreader = csv.DictReader(csvfile)
                                 headers = taskreader.fieldnames
-                                last_task = OrderedDict()
                                 for row in taskreader:
                                     last_task = row
                                     lines.append(row)
